+++ pages/index.tsx
@@ -1,10 +1,27 @@
-import { useJsApiLoader } from '@react-google-maps/api';
+import { ErrorMessage } from '@hookform/error-message';
+import { useJsApiLoader, GoogleMap, Marker } from '@react-google-maps/api';
 import React, { ReactElement, useState } from 'react';
-import { SubmitHandler, useForm } from 'react-hook-form';
-import { Button, Notification } from 'react-bulma-components';
+import {
+  Button,
+  Notification,
+  Container,
+  Section,
+  Form,
+  Columns,
+  Loader,
+} from 'react-bulma-components';
+import { Controller, SubmitHandler, useForm } from 'react-hook-form';
+
 import { Meta, Layout } from '../components';
-import Main from '../components/pages/Main';
 import { LatLng, SelectedIdols, NotificationToast } from '../types/Type';
+import idols from '../utils/idols.json';
+
+const { Field, Select, Label, Help, Control: BulmaControl } = Form;
+
+const containerStyle = {
+  width: '100%',
+  height: '60vh',
+};
 
 const Home = (): ReactElement => {
   const [pinLatLng, setPinLatLng] = useState<LatLng | undefined>();
@@ -22,34 +39,31 @@ const Home = (): ReactElement => {
     googleMapsApiKey: process.env.NEXT_PUBLIC_GOOGLE_MAPS_API_KEY,
   });
 
-  const { control, register, handleSubmit } = useForm<SelectedIdols>();
+  const {
+    control,
+    handleSubmit,
+    formState: { errors },
+    watch,
+  } = useForm<SelectedIdols>();
 
   const initPosition: LatLng = {
     lat: 35.69575,
     lng: 139.77521,
   };
 
+  const isUnique = () => {
+    const idols = [watch('idol1'), watch('idol2'), watch('idol3')];
+    if (idols[0] !== idols[1] && idols[1] !== idols[2] && idols[2] !== idols[0]) {
+      return true;
+    } else {
+      return '同じアイドルを選択しています';
+    }
+  };
+
   const onSubmit: SubmitHandler<SelectedIdols> = async (values) => {
     setNotification({ ...notification, show: false });
     const idols = [values.idol1, values.idol2, values.idol3];
 
-    if (idols[0] == idols[1] || idols[1] == idols[2] || idols[2] == idols[0]) {
-      setNotification({
-        show: true,
-        type: 'warning',
-        body: '3人別々のアイドルを選択してください',
-      });
-      return;
-    }
-
-    if (idols.includes('')) {
-      setNotification({
-        show: true,
-        type: 'warning',
-        body: 'アイドルを選択してください',
-      });
-      return;
-    }
     setLoading(true);
     await fetch('/api/idol2map', {
       method: 'POST',
@@ -93,16 +107,120 @@ const Home = (): ReactElement => {
             <Button remove onClick={onNotificationClose} />
           </Notification>
         )}
-        <Main
-          pinLatLng={pinLatLng}
-          initPosition={initPosition}
-          isLoaded={isLoaded}
-          loading={loading}
-          register={register}
-          control={control}
-          handleSubmit={handleSubmit}
-          onSubmit={onSubmit}
-        />
+        <Container>
+          <Section>
+            <form onSubmit={handleSubmit(onSubmit)}>
+              <Columns>
+                <Columns.Column>
+                  <Field>
+                    <Label>アイドル</Label>
+                    <BulmaControl>
+                      <Controller
+                        name="idol1"
+                        control={control}
+                        rules={{ required: 'アイドルを選択してください', validate: isUnique }}
+                        render={({ field }) => (
+                          <Select color="info" {...field}>
+                            {idols.idols.map((idol) => (
+                              <option key={idol.id} value={idol.label}>
+                                {idol.label}
+                              </option>
+                            ))}
+                          </Select>
+                        )}
+                      />
+                    </BulmaControl>
+                    <ErrorMessage
+                      errors={errors}
+                      name="idol1"
+                      render={({ message }) => <Help color="danger">{message}</Help>}
+                    />
+                  </Field>
+                </Columns.Column>
+
+                <Columns.Column>
+                  <Field>
+                    <Label>アイドル</Label>
+                    <BulmaControl>
+                      <Controller
+                        name="idol2"
+                        control={control}
+                        rules={{ required: 'アイドルを選択してください', validate: isUnique }}
+                        render={({ field }) => (
+                          <Select color="info" {...field}>
+                            {idols.idols.map((idol) => (
+                              <option key={idol.id} value={idol.label}>
+                                {idol.label}
+                              </option>
+                            ))}
+                          </Select>
+                        )}
+                      />
+                    </BulmaControl>
+                    <ErrorMessage
+                      errors={errors}
+                      name="idol2"
+                      render={({ message }) => <Help color="danger">{message}</Help>}
+                    />
+                  </Field>
+                </Columns.Column>
+
+                <Columns.Column>
+                  <Field>
+                    <Label>アイドル</Label>
+                    <BulmaControl>
+                      <Controller
+                        name="idol3"
+                        control={control}
+                        rules={{ required: 'アイドルを選択してください', validate: isUnique }}
+                        render={({ field }) => (
+                          <Select color="info" {...field}>
+                            {idols.idols.map((idol) => (
+                              <option key={idol.id} value={idol.label}>
+                                {idol.label}
+                              </option>
+                            ))}
+                          </Select>
+                        )}
+                      />
+                    </BulmaControl>
+                    <ErrorMessage
+                      errors={errors}
+                      name="idol3"
+                      render={({ message }) => <Help color="danger">{message}</Help>}
+                    />
+                  </Field>
+                </Columns.Column>
+              </Columns>
+
+              <Button.Group>
+                <Field kind="group">
+                  <BulmaControl>
+                    <Button color="primary" disabled={loading} className="has-text-white">
+                      What3Idols!!!
+                    </Button>
+                    {loading ? <Loader /> : null}
+                  </BulmaControl>
+                </Field>
+              </Button.Group>
+            </form>
+          </Section>
+
+          <Section>
+            {isLoaded && (
+              <GoogleMap
+                mapContainerStyle={containerStyle}
+                zoom={10}
+                center={pinLatLng || initPosition}
+                options={{
+                  mapId: process.env.GOOGLE_MAPS_MAP_ID,
+                }}
+              >
+                <Marker position={pinLatLng} />
+              </GoogleMap>
+            )}
+          </Section>
+        </Container>
       </Layout>
     </>
   );
